fix: introduce WorkManager based SyncRequestService
This commit introduces a new implementation of the SyncRequestService
that depends on the Android WorkManager to schedule and execute jobs.

This will allow us to continue background work in the sync queue even
after the application is in the background, without causing crashes
on Android >= 8 (as the old SyncScheduler does with a background service).

depends on: https://github.com/wireapp/wire-android-sync-engine/pull/446

fixes: https://github.com/wireapp/android-project/issues/314